ft_printf

+++ ft_printf/linux/srcs/ft_hexadec_handle.c
@@ -6,7 +6,7 @@
 /*   By: lchantel <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/19 17:50:52 by lchantel          #+#    #+#             */
-/*   Updated: 2020/07/28 00:41:24 by lchantel         ###   ########.fr       */
+/*   Updated: 2020/07/31 00:30:59 by lchantel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -79,7 +79,7 @@ int			ft_hexadec_handle(char **space_info, unsigned long long value,
 	sp_info[3] = (sp_info[1] - sp_info[2] > 0)
 	? sp_info[0] - sp_info[1] : sp_info[0] - sp_info[2];
 	sp_info[3] -= (**(space_info + 1) == 'p' && value) ? 2 : 0;
-	sp_info[3] += (!value && **(space_info + 2) == '0'
+	sp_info[3] += (!(int)value && **(space_info + 2) == '0'
 	&& **(space_info + 1) != 'p') ? 1 : 0;
 	str_value[1] = str_value[0];
 	*len_res = ft_stdout_hex(space_char, sp_info, str_value, space_info);
